src/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "twophase"
version = "0.1.0"
description = "Two-phase flow solver (CLS + CCD + Projection method)"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.11",
]

[project.optional-dependencies]
gpu = ["cupy-cuda12x"]
dev = ["pytest>=7", "pytest-cov"]

[tool.setuptools.packages.find]
where = ["."]
include = ["twophase*"]

[tool.pytest.ini_options]
testpaths = ["twophase/tests"]
